resolve unspecified addresses
this fixes a tricky bug. unspecified (0.0.0.0) addresses are
not meant to be part of connections. libp2p assumes that conns
have specified addresses. 0.0.0.0 addrs from listeners need to
be resolved when assigned to a connection.

Im not sure how transports like TCP achieve this (is this in manet?
default from golang/net package?), but there is a difference between
QUIC and TCP. For now, im using the addrutil function, but this is
probably not the right thing to do.